requirements.txt
fastapi
uvicorn
pydantic
jsonschema
pyyaml
dicttoxml
structlog>=23.0.0
python-dotenv>=1.0.0
